+++ DivarClone/DivarClone/urls.py
@@ -18,9 +18,18 @@ from django.contrib import admin
 from django.urls import include, path
 
 
+from rest_framework_simplejwt.views import (
+    TokenObtainPairView,
+    TokenRefreshView,
+)
+
+
 urlpatterns = [
     path("ads/", include("ads.urls")),
     path('admin/', admin.site.urls),
     path('api/', include('ads.api.urls', namespace='api')),
+    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
+    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
+    path('api/users', include(('users.urls', 'users') , namespace = 'users')),
 
 ]
